# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Campus Marketplace API
  version: 1.0.0
  description: 校园二手交易与闲置物品流转平台 API
servers:
  - url: http://localhost:7001
paths:
...
  /api/orders:
    get:
      operationId: listMyOrders
      summary: 获取我的订单
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: status
          in: query
          schema:
            type: string
            enum: [pending_payment, paid, completed, cancelled]
      responses:
        "200":
          description: 我的订单列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderPage"
        "401":
          description: 未认证
    post:
      operationId: createOrder
      summary: 发起购买
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: 订单创建成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Order"
        "400":
          description: 库存不足或商品不可购买
        "401":
          description: 未认证
        "404":
          description: 商品不存在
  /api/orders/{id}:
    get:
      operationId: getOrder
      summary: 获取订单详情
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 订单详情
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Order"
        "403":
          description: 无权查看
        "404":
          description: 订单不存在
  /api/orders/{id}/pay:
    post:
      operationId: payOrder
      summary: 模拟支付
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 支付成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Order"
        "400":
          description: 订单状态不允许支付
        "403":
          description: 无权操作
  /api/orders/{id}/confirm:
    post:
      operationId: confirmOrder
      summary: 确认收货
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 确认成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Order"
        "400":
          description: 订单状态不允许确认
        "403":
          description: 无权操作
  /api/orders/{id}/cancel:
    post:
      operationId: cancelOrder
      summary: 取消订单
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 取消成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Order"
        "400":
          description: 订单状态不允许取消
        "403":
          description: 无权操作
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Item:
      type: object
      required:
        [
          id,
          sellerId,
          categoryId,
          title,
          price,
          quantity,
          availableQuantity,
          status,
          images,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        categoryId:
          type: integer
        title:
          type: string
        description:
          type: [string, "null"]
        price:
          type: number
          minimum: 0
        quantity:
          type: integer
          minimum: 0
        availableQuantity:
          type: integer
          minimum: 0
        status:
          type: string
          enum: [pending_review, listed, reserved, sold, rejected, removed]
        rejectReason:
          type: [string, "null"]
        images:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        seller:
          type: object
          properties:
            id:
              type: integer
            username:
              type: string
            studentId:
              type: string
        category:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
...
    Order:
      type: object
      required:
        [
          id,
          orderNo,
          buyerId,
          itemId,
          quantity,
          totalPrice,
          status,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: integer
        orderNo:
          type: string
        buyerId:
          type: integer
        itemId:
          type: integer
        quantity:
          type: integer
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending_payment, paid, completed, cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        item:
          $ref: "#/components/schemas/Item"
    CreateOrderInput:
      type: object
      additionalProperties: false
      required: [itemId]
      properties:
        itemId:
          type: integer
        quantity:
          type: integer
          minimum: 1
          default: 1
    OrderPage:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
...